io_u: residiual size should be unsigned long long
If we target a number of bytes of IO and keep subtracting a
xfer_buflen, we need to ensure the residual length is of the
same size.

+++ io_u.h
@@ -75,7 +75,7 @@ struct io_u {
 
 	struct io_piece *ipo;
 
-	unsigned int resid;
+	unsigned long long resid;
 	unsigned int error;
 
 	/*
